+++ src/extension.ts
@@ -1,6 +1,6 @@
 "use strict";
 
-import { window, ExtensionContext } from "vscode";
+import { window, ExtensionContext, commands, workspace } from "vscode";
 import Sundial from "./sundial";
 
 const sundial = new Sundial();
@@ -12,6 +12,43 @@ export function activate(context: ExtensionContext) {
   context.subscriptions.push(window.onDidChangeWindowState(check));
   context.subscriptions.push(window.onDidChangeActiveTextEditor(check));
   context.subscriptions.push(window.onDidChangeTextEditorViewColumn(check));
+
+  commands.registerCommand("sundial.switchToNightTheme", async () => {
+    console.info(`Switching to your night theme...`);
+    await sundial.updateConfig();
+    await sundial.disablePolos();
+    sundial.changeThemeTo(sundial.SundialConfig.nightTheme);
+  });
+
+  commands.registerCommand("sundial.switchToDayTheme", async () => {
+    console.info(`Switching to your day theme...`);
+    await sundial.updateConfig();
+    await sundial.disablePolos();
+    sundial.changeThemeTo(sundial.SundialConfig.dayTheme);
+  });
+
+  commands.registerCommand("sundial.toggleDayNightTheme", async () => {
+    console.info(`Toggling your theme now...`);
+    await sundial.updateConfig();
+    await sundial.disablePolos();
+    const currentTheme = sundial.WorkbenchConfig.colorTheme;
+    if (currentTheme === sundial.SundialConfig.dayTheme) {
+      sundial.changeThemeTo(sundial.SundialConfig.nightTheme);
+    } else if (currentTheme === sundial.SundialConfig.nightTheme) {
+      sundial.changeThemeTo(sundial.SundialConfig.dayTheme);
+    } else {
+      window.showInformationMessage(
+        "Toggling not working, please set your theme to one of your configurated sundial themes!"
+      );
+    }
+  });
+
+  commands.registerCommand("sundial.continueAutomation", async () => {
+    console.info(`Attaching the polos to the sundial again...`);
+    await sundial.updateConfig();
+    sundial.automater();
+    sundial.polos = true;
+  });
 
   if (sundial.SundialConfig.interval !== 0) {
     sundial.automater();